native: remove deprecated API
The former so-called "native" API from the 2.x architecture is
replaced by the new "alchemy" API, which rebases the existing native
API services over the Copperplate library in userland.

Now that alchemy is close to completion, and we can resolve all the
references to the old native API over the new library, we don't need
the former in the tree anymore.