copperplate/heapobj-shared: sanitize heap creation, fix extension
We still do not allow the kernel to remap a memory mapping to a
different base address, so the heap extension feature is likely to be
unreliable because of random failure to remap at the same place. So
either we allow MREMAP_MAYMOVE, which would not be a trivial task in
terms of multi-thread/multi-process synchronization, or we just get
rid of the feature.

This needs more thoughts.