copperplate/clockobj, lib: assume fixed clock resolution during runtime
For performance reason, we want to run locklessly for common time unit
conversions, so the clockobj implementation now assumes that the clock
resolution will not be updated on-the-fly.

This means that clockobj_set_resolution() shall be called from init
- or known to be safe - context only.

COPPERPLATE_PROTECT sections have been removed around client API calls
which do not involve locking anymore.