copperplate/clockobj: clockobj_ticks_to_clock() -> clockobj_convert_clocks()
We rename this conversion call to fix a misnomer, taking a timespec
for input instead of a count of ticks, since most callers will want to
pass this.

CAUTION: The input value shall be based on CLOCK_COPPERPLATE, so this
is really a conversion routine from CLOCK_COPPERPLATE to other clock.