cobalt, copperplate/syncobj: introduce fast internal events
Condvars are the basic mechanism used internally by Copperplate's
syncobj for implementing resource waits, so this ought to be fast and
optimized.

We introduce a streamlined internal version of Cobalt's condvar
support, and use it in Copperplate. This version removes all
extraneous code which is either redundant or useless when called from
the Copperplate innards.