copperplate/heapobj, vxworks: sanitize locking, simplify wrapping
We remove the per-heap object lock, which was duplicating the outer
locking from a sync object the heap belongs to in most cases, fixing
the only caller who did expect the heap object to handle locking
internally (i.e. VxWorks's memParLib).

Secondly, we get rid of the indirection to invoke the allocator
routines via a call table in the pshared case, which brought
absolutely zero value, but overhead. At this chance, we also inline
most of the calls to TLSF/malloc for private allocation.