nucleus/pod: dequeue kicked threads from any synch wait list
Threads pending a linux signal on entry to xnsynch_acquire/sleep_on()
will be prevented from waiting in xnpod_suspend_thread() as XNKICK is
raised for them, and will return immediately to the caller.

In such a case, we have to call xnsynch_forget_sleeper() for them, to
update the synch wait list, preventing stale linkage.