cobalt/monitor: plug signal leakage on entry to wait op
A thread about to wait for a GRANT/DRAIN signal might have sent a
signal prior to running the syscall (usually the converse one). We
must wake any satisfied sleeper prior to entering sleep.

In additional, the wakeup code for the DRAIN signal has been fixed
(inverted test of the drain wait list).