lib/cobalt: make ownership check unconditional on unlocking
f1a968696 reduced the amount of runtime checks done for non
ERRORCHECK mutexes, to speed up operations in known-to-be-safe
cases. Unfortunately, several APIs want to combine recursive mutex and
check for lock imbalance upon unlocking, which puts them at odds with
this change.

We introduce back the unconditional check for lock imbalance in
__wrap_pthread_mutex_unlock() to fix this issue, without resorting to
significant overhead in client APIs.