copperplate/threadobj, lib: fix locking rules of threadobj_set_priority()
Since threadobj_set_priority() must receive the thread lock on entry,
prior to dropping it before applying the priority change via a
syscall, it might be unable to lock it back before returning to the
caller, in case the thread has exited in the meantime.

Instead of attempting to re-lock the object eventually - bluntly
ignoring the operation status - we should rather behave fully as a
lock-dropping call and document it accordingly. Callers should grab
that lock again in case more changes are required (which none of them
currently want anyway).