copperplate/syncobj, lib: fix multiple races caused by deletion
This patch is a major rework of the syncobj deletion code. Basically,
deleting a synchronization object under heavy use just did not work,
triggering a variety of deadly races.

At this chance, we also fixed several syncobj API services (name
and/or goal), to fit the GRANT/DRAIN semantics introduced by the
Cobalt monitors (but still relevant for Mercury nevertheless).