kernel/cobalt: gate lock must be held to destroy a monitor
We now require the caller of the monitor deletion service to hold the
relevant gate lock. This forcibly plugs a common race in userland APIs
based on monitors, allowing them to request a deletion atomically.

As a side effect, this reserves the monitor deletion service to
Xenomai threads, excluding regular Linux tasks. Copperplate, which is
the only and main user of Cobalt monitors does already require this
internally anyway.