hal: start removing the HAL interface
Rationale: the HAL was introduced to support multiple real-time
enablers by design, which is not relevant anymore. Besides, the
interrupt pipeline is in itself an abstraction layer, hiding the
architecture details from the client code.

We start getting rid of this useless indirection, moving most of its
contents to the system/xnarch layer.