hal: remove last references to CPU frequency
Now that all architectures implement the sysinfo v2 format including
x86 (ipipe), we may safely remove all references to the CPU frequency,
which used to be a wrong placeholder for the clock frequency inherited
from the x86 dark ages.

We only need to know about the timer and clock devices frequencies,
for timing duties.