nucleus: remove sysheap_size module parameter
Nobody ever heard of this parameter, and the preferred way has always
been to set CONFIG_XENO_OPT_SYS_HEAPSZ for defining the system heap
size instead. This value is not something one would change for a quick
check on the fly anyway.

This way, we can always introduce build time computations based on
this value.