nucleus: use builtin shadow migration service
The latest pipeline core introduces __ipipe_migrate_head(), to start
the migration of the current task to the head domain. To complete the
migration duties, ipipe_migration_hook() as implemented by the nucleus
is invoked. This mechanism does not require any third-party helper to
complete the migration, i.e. there is no gatekeeper thread anymore.

This patch rebases xnshadow_harden() over this new feature, and also
provides the former gatekeeper-based implementation in the legacy
support code for older pipelines.