cobalt: introduce user visible state window
We can't reuse the common thread state flags to hold the monitor grant
condition (former XNGRANT), since the u_mode area should be altered by
the owner thread only, unless all accesses are changed to atomic ops,
which would be overkill.

To fix this, we replace the former u_mode data by a more generic user
visible structure shared between the nucleus and the Cobalt libraries,
in which the copy of the kernel thread state bits is stored separately
from the monitor grant flag.