posix: fix linux-side posix threads priority
xnshadow_map expects the thread being mapped to already have the
correct priority and policy on linux side if this thread does not
use the completion mechanism, in order to avoid sending a signal
which would cause a switch to secondary mode.

However, user-space posix skins start with the same priority as
their parent thread. The result is that, on Linux side, they keep
that wrong priority until their priority is synchronized for whatever
reason.

Fix this by explicitely changing priority on Linux side before emitting
the syscall that will result in a call to xnshadow_map.