cobalt/event, lib/cobalt: introduce support for event flag groups
An event flag group is a synchronization object represented by a
long-word structure; every available bit in such word can be used to
map a user-defined event flag.  When a flag is set, the associated
event is said to have occurred.

Xenomai threads and interrupt handlers can use event flags to signal
the occurrence of events to other threads; those threads can either
wait for the events to occur in a conjunctive manner (all awaited
events must have occurred to wake up), or in a disjunctive way (at
least one of the awaited events must have occurred to wake up).

We expose this non-POSIX feature through the internal API, as a fast
IPC mechanism available to the Copperplate interface.