copperplate: synchronize threadobj_start() with target thread
A thread started with high priority should enter the user code before
threadobj_start() returns to a low priority caller. IOW, we must avoid
priority inversions due to some resources not being immediately
available to the started thread when it executes the internal
prologue, causing such thread to block or be demoted priority-wise.

This is particularly required when running over the Cobalt core, when
the started thread has to migrate to secondary mode for running
portions of the prologue code, in which case its priority is
automatically demoted.

This patch introduces the required synchronization between the
starting thread and the threadobj_start() caller, only when the latter
has a strictly lower priority than the former. In all other cases, the
thread issuing threadobj_start() may keep running without
synchronizing.

NOTE: this mechanism makes most sense in uniprocessor configurations,
where assumptions could be made on the effect of the respective
priorities of both the started and starter thread. I.e. the starter
may assume the started thread has entered its processing routine
before threadobj_start() returns, for any relevant purpose. However,
we enable it for SMP configurations as well, to keep these assumptions
valid regardless.