lib/testsuite: fix expected execution sequence for timer-related tests
First call to timerobj_init() may trigger a switch to secondary mode
over Cobalt, as we create the timer server thread internally. For this
reason, we can't expect the related tests to keep running in primary
mode unaffected when traversing this code.

Fix all expected execution sequences accordingly in the various tests
involved.