nucleus: deprecate priority coupling
This feature does not fully deliver the initial promise of boosting
the linux kernel priority in every circumstance requiring it (granted,
only in a few uncommon cases, but still).

However, it still has a measurable impact on latency during primary
<-> secondary mode switches, although some progress has been made in
this area. Besides, it became overly complex over time to maintain, as
more issues where discovered and fixed, compared to the fairly small
number of use cases which might actually benefit from it.

For these reasons, we have already pulled this feature out from the
Xenomai 3 core. It is now marked as deprecated in 2.6.x as well and
defaults to disabled, to give a clear hint that applications should
not depend on it.

The short version is: 'Darn, I was wrong'.