x86: fix mayday support in 32bit+SEP mode
The mayday syscall (__xn_sys_mayday) will not return to its caller,
but rather to the runaway code originally preempted by the mayday
event, once the faulty thread is relaxed.

For this reason, we have to fix up the user stack pointer before
returning from the mayday syscall handler, to discard the data pushed
onto the stack by the sysenter prologue in vDSO to kick
__xn_sys_mayday.

This bug is only seen on x86_32 when SYSENTER is enabled, because we
won't get any data pushed onto the user stack for issuing
__xn_sys_mayday in long mode, or whenever a legacy int80 syscall is
used.

We do the fix up simply by restoring the original ESP before returning
from the mayday syscall handler, which will do the right thing when
SEP is enabled, or lead to a nop otherwise.