cobalt/drivers/serial: introduce driver for IMX SOCs UARTs
This RTserial driver is for the Freescale IMX internal UART compatible
controllers. It has been tested with Freescale's 3.0.15 kernel on an
i.MX53 and i.MX6Q board but it should work with more recent mainline
kernel versions as well. The modem control and line state is handled
similar to the 16550A UART, if possible. But this part is still weak
and not very well tested.

Signed-off-by: Wolfgang Grandegger <wg@denx.de>