hal/x86: dynamic tsc detection again
A simpler use of the I-pipe core for 3.4 dynamic tsc detection feature.
When compiled with CONFIG_X86_TSC and running on a machine without a tsc,
the I-pipe core for 3.4 will automatically use the 8254 emulation and pass
Xenomai the correct value of the clock frequency. What is only left to be done
is:
- display the correct clockdev in /proc/xenomai/timer
- do not arm the tsc feature when the user-space test for compatible features.

As Xenomai was not even booting in this configuration prior to 3.4, do this
unconditionally.

When configured without CONFIG_X86_TSC and running on a system with a tsc, no
attempt is made to use the tsc whereas the I-pipe core is able to do it (this is
the difference with commit 89cb4933ec93d4e6209e2f681d1f932467d98db6).