nucleus/shadow: do not clear PIP boost for an exiting thread
When a shadow thread is running the linux-side task exit code, we must
not attempt to de-boost it, in case it holds a contended PIP-managed
synchronization object.

Not only resetting the priority of a terminating thread is useless,
but since that thread will clear its own struct_task backlink _before_
de-boosting itself, there is no way to send and process a lostage
renice request for it.

We now check for the zombie state of the de-boosted thread, to prevent
priority update if set.