testsuite/xeno-test-run: stop abusing the shell
xeno-test-run communicates with the xeno-test script by using two pipes
which file descriptors are left open when executing a shell. This does not
work with newer versions of busybox, and thinking about it, we should not
have done this anyway.

So, create two fifos instead of the two pipes, and let the child shell of
xeno-test-run use them by usual means.