cobalt/kernel/intr: Convert intrlock to a sleeping Linux lock
Conceptually, all users of this lock are supposed to run over Linux
context already. Moreover, latest I-pipe patches complain that
ipipe_virtualize_irq is called with the Xenomai domain stalled which
invalidates that useful internal bug check.

So rework the locking for dumping scheduling statistics and convert
intrlock to a sleeping Linux variant to avoid the alarm.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>