cobalt: remove HAL interface
Originally, the HAL interface was aimed at abstracting the real-time
enabler on top of which the nucleus sits, when running over the linux
kernel. Along the years, we only used the Adeos pipeline for this
purpose, and 3.x is writing this dependency in stone for the Cobalt
core.

Therefore the HAL layer becomes pointless, and removing it simplifies
a great deal of low level Xenomai code, making architecture ports even
simpler.

In the same move, we dismiss support for legacy pipelines. A pipeline
from the recent "core" series, exposing API revision #2 or better is
required to build and run Xenomai 3.x.

NOTE: nios2 and sh4 are not ready for prime time, since we don't have
a core pipeline implementation for them yet. We leave them in tree
though, waiting for catch-up.