cobalt: move kernel-based Xenomai threads over native linux kthreads
This is a long due change aimed at stopping the absolute non-sense of
creating and managing non-linux threads in a linux kernel.

Aside of a massive simplification of the low-level code, using native
linux kthreads to support kernel-based Xenomai threads also gets rid
of the braindamage non-linux stack contexts, which is quite of a
relief.

In addition, this introduces the secondary runtime mode for Xenomai
kthreads, despite we don't expose it though any official interface for
sanity reasons (i.e. we have no syscall interface to do the required
fixups in case of mode mixing, so we don't want to open this feature
broadly).