copperplate/timerobj: unify implementation on internal timer server
Using SIGEV_THREAD over the Mercury core is way overkill, with a
short-lived thread being created at each timer expiration, for running
the handler. Besides, this support is known to be broken for glibc 2.4
-> 2.9 at least:
http://sourceware.org/bugzilla/show_bug.cgi?id=7094

We unify the timer management over the Cobalt approach, managing our
own internal timer server. This also means that user handlers should
not behave as CPU hogs since delivery is single-threaded, but since we
are emulating ISR contexts RTOS-wise, they should already honor this
requirement.

At the moment, outstanding timers are managed over a single linked
list, which is fine for up to ~10 instances. We plan to provide for a
binary heap later, for handling larger configurations.