psos/init: time_slice is specified in ticks instead of ns
The pSOS skin uses a default timeslice of 1000 sec.  In comments it is
mentioned that the value is specified in ms while the place where the
variable is used assumes that it is in ticks.

Rename time_slice to time_slice_in_ticks to make this clear.
Additionally, increase the timeslice from the intended 1ms to 5ms to
reduce the overhead a bit.

Signed-off-by: Ronny Meeus <ronny.meeus@gmail.com>