rtcan_module: Fix tx_timeout buffer size inconsistency
When formatting /proc/rtcan/sockets, tx_timeout is defined as
char[16], but a length of 20 is passed to rtcan_get_timeout_name. If
tx_timeout for any socket is infinite, this causes snprintf to zero
outside the buffer, leading to stack overflow and a kernel panic.
Make the tx_timeout buffer 20 bytes long instead.

Signed-off-by: Paul Janzen <pcj@xenomai.sez.to>
Acked-by: Wolfgang Grandegger <wg@grandegger.com>