build: introduce core API revision level in userland
This changeset provides the CONFIG_XENO_UAPI_LEVEL symbol to userland
applications, defining the current core API revision level being
implemented in user-space. This level can be seen as the current
version of the Copperplate library, it is stored in config/apirev.

At this chance, most of the version information has been centralized
in config/version and config/apirev, at the notable exception of the
kernel ABI revision level, which is Cobalt-specific and therefore
remains available from the per-arch asm/features.h header file.

CAUTION: config/version and config/apirev are parsed when the autoconf
system is bootstrapped using scripts/bootstrap, which means that any
change to the version information requires a bootstrap to be
effective.  See configure.in for details on the config/version format.