x86/hal: allow SMI control at run-time
Having to recompile the kernel when we want to change the settings regarding
SMI workaround is not really compatible with distributing pre-compiled kernels.
So we make the control possible with two kernel parameters. Allowing to compile
out the functionality did not save a lot of room anyway.

At this chance, suppress the PCI ids table which was a maintenance burden, any
Intel chipset with a PCI-ISA bridge is assumed to allow SMI workaround.

By default, the module detects the chipset but does nothing. Setting:
- xeno_hal.smi to -1 disables SMI detection
- xeno_hal.smi to 1 enables SMI workaround, the parameter xeno_hal.smi_mask
controls which bits are masked. If not set, the bit 0 is masked which means
SMI are globally disabled.