testsuite/switchtest: handle device open error
Currently, switchtest uses the failure of the first ioctl on its device to
know that it was in fact not the correct device. But the error can also be
due to the ioctl failing to allocate memory with kmalloc, which indeed happens
on system with many CPUs. Fix this issue by:
- using vmalloc instead of kmalloc to have a higher upper bound on allocation
size;
- only considering ENOSYS and ENOTTY as indication that the device is not the
right one.