lib/cobalt: allow compiling applications with _FORTIFY_SOURCE
When _FORTIFY_SOURCE is set, some services are replaced at compilation
time with versions performing additional checks on their
arguments. This is a problem if the same services need to be wrapped
by Xenomai: the checked versions are not wrapped, causing switches to
secondary mode.

This commit provides wrappers for the checked services.

The wrappers simply call the non checked version wrapper: calling
glibc checked services directly would break compilation with libcs
without this option, meaning old glibcs and alternative libcs.

In order to enable the checks provided by _FORTIFY_SOURCE, the Xenomai
binaries should be compiled with this option.

We attempt to auto-detect whether the underlying *libc supports
_FORTIFY_SOURCE at configuration time. --enable-fortify is introduced
to enable or disable the corresponding support in Xenomai libs.

If the support is enabled, checked versions of the Cobalt services are
redirected to checked versions of the *libc services. Otherwise, if a
checked version of a Cobalt service ends up being called, a fatal
error is raised.