posix: allow compiling applications with _FORTIFY_SOURCE
When _FORTIFY_SOURCE is set, some services are replaced at compilation time
with versions performing additional checks on their arguments. This is a problem
if the same services need to be wrapped by Xenomai: the checked versions are
not wrapped, causing switches to secondary mode.

This commit provides wrappers for the checked services.

The wrappers simply call the non checked version wrapper: calling glibc checked
services directly would break compilation with libcs without this option, meaning
old glibcs and alternative libcs.

In order to enable the checks provided by _FORTIFY_SOURCE, Xenomai itself should
be compiled with this option.