posix: _FORTIFY_SOURCE support take #2
Add a test to configure to detect whether the underlying libc supports
_FORTIFY_SOURCE and an option allowing to enable or disable the corresponding
support in Xeonmai libs.

If the support is enabled in Xenomai libs, checked versions of the posix skin
services are redirected to checked versions of the glibc services, if not and
a checked version of a posix skin service ends up being called, print an error
message and exit.