Invoke mlockall on every skin library initialization
The nucleus refuses to create threads if userspace didn't lock its
memory, that's the rule since day #1 basically. So far we require
userspace to do this explicitly, but there is actually nothing to lose
by doing it automatically during the initialization of a skin library.
In special scenarios, userspace could still temporarily unlock the
memory again.

As we lock unconditionally, there is no more need to lock/unlock around
POSIX, PSOS+ and uITRON auto-shadowing.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>