posix: Control auto-shadowing via environment variable
There is no need anymore to disable auto-shadowing when dlopen support
is enabled as we preserver the scheduling parameters of the context.

Instead, port the environment variable based control from -forge to 2.6:
Only if XENO_NOSHADOW is set to a non-empty string, we refrain from
shadowing the caller of the POSIX skin init function.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>