copperplate/threadobj: prepare for exporting thread state via sysregd
In order to export copperplate threads via the sysregd interface, we
need more information to be returned by threadobj_stat(), such as the
current timeout value and CPU number, for both cores.

In addition, more run state information is exposed to determine
whether a thread is currently blocked, with or w/o a timeout.

Finally, the costly thread suspend/resume and syncobj wait hooks are
removed, since we can retrieve a thread state on demand, instead of
tracking state changes dynamically.