alchemy/task: introduce rt_task_set_affinity(), remove T_CPU() macro
The former T_CPU() macro would only allow for CPUs #0-7 to be present
in the affinity mask, which is too restrictive, particularly over the
Mercury core.

We introduce rt_task_set_affinity() for the purpose of setting the CPU
affinity of a task, which takes a cpu_set_t argument as required for
high end hardware configurations.