Work around gcc-4.6 bug
This avoid that gcc bug 56712 lets the initialization of the native skin
fail. See http://gcc.gnu.org/bugzilla/show_bug.cgi?id=56712 for the
background. Affects in particular Ubuntu 12.04 LTS.

The current code only requires the noclone attribute as workaround for
the native skin. But we are better conservative and add it to all
constructors in case their structure changes in an unfortunate way in
the future.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>