nucleus/sem, cond: allow reinitializing anon shared objects
We allow reinitializing shared anonymous semaphores and condvars.

Rationale: if the process creating such object exits, we may assume
that other processes sharing that object won't be able to keep on
running.

This avoids the requirement for applications running in pshared mode
to destroy these synchronization objects explicitly before exiting.