make rtcan_flexcan driver compatible with non OpenFirmware platforms
This also fixes the rx fifo global mask for mx6:

The default rx fifo global mask register, newly introduced in mx6q,
is 0xffffffff and the reset value in Message buffers(can be reused as the
memory of rx fifo filter table) is none zero, it will wrongly cause the
can to be unable to recevie packets due to filter.
We need to clear it to make sure to receive all packets.

Signed-off-by: Henri Roosen <henriroosen@gmail.com>
Acked-by: Wolfgang Grandegger <wg@grandegger.com>