cobalt/arm: be flexible with selecting unlocked context switch on SMP
Starting with Linux 3.8, due to the change in the ASID roll-over
mechanism, we can disable unlocked context switch on SMP platforms,
which improves user-space scheduling latency and may provide a more
interesting trade-off especially since SMP platforms do not flush the
cache on mm switch, so have short mm switch time.

So, replace the Kconfig check that was forcing unlocked context switch
on SMP, with a compile-time check, which takes the kernel version into
account.