cobalt/kernel/sched: pre-calculate weighted thread priority
Weighted priority is a per-thread priority level which includes the
weight value of the current scheduling class for the thread
being considered.

This patch introduces pre-calculation of this value when a thread
priority or scheduling class changes, which happens much less
frequently in a thread's lifetime than uses of such value, typically
for enforcing PIP on synchronization objects.