include/cobalt/kernel: drop obsolete queue manipulation helpers (queue.h)
We have no more users for legacy queues. From now on, new code should
use regular kernel lists instead.

Former queue debugging features are dropped in the same move. For the
time being, it is assumed that enabling CONFIG_DEBUG_LIST should
provide for sufficient test coverage, even if the duplicate insertion
and spurious deletion cases might not be systematically and
specifically detected as they used to be.