include/cobalt: fix build over uClibc
Cobalt does implement pthread_mutexattr_set/getprotocol(), as well as
pthread_condattr_set/getclock(), so these may be declared
unconditionally in a Cobalt-specific header file.

In addition, we detect when pthread_yield() is a macro such as in the
uClibc case, so prevent preprocessing errors with
COBALT_DECL(). uClibc wraps pthread_yield() to sched_yield(), which we
override too, so that's fine with us.