build: introduce symbols-only debug level
--enable-debug[=partial]

"symbols" enables debug symbols to be compiled in the libraries and
executables, still turning on the optimizer (-O2). This option has no
overhead, it is useful to get meaningful backtraces using gdb while
running the application at nominal speed.

"partial" includes "symbols", and also turns on internal consistency
checks within the Xenomai code (mostly present in the Copperplate
layer). The __XENO_DEBUG__ macro is defined, for both the Xenomai core
and the applications getting their C compilation flags from the
xeno-config script (i.e. xeno-config --cflags). The partial debug mode
implicitly turns on --enable-assert. A measurable overhead is
introduced by this level.  This is the default level when
--enable-debug is mentioned with no level specification.

"full" includes "partial" settings, but the optimizer is disabled
(-O0), and even more consistency checks may be performed.  In addition
to __XENO_DEBUG__, the macro __XENO_DEBUG_FULL__ is defined. This
level introduces the most overhead, which may triple the worst-case
latency, or even more.