cobalt/posix/thread: sanitize mapping from core TCB to cobalt/posix thread
Testing the personality to determine whether a thread originates from
cobalt/posix is wrong, since that information may change dynamically
as extra personalities stack over the base posix one.

In addition, we may assume that any userland thread with a valid
Xenomai shadow extension which switches to kernel mode, is originally
a cobalt/posix thread, spawned from lib/cobalt.

As a consequence of this, thread2pthread() is dropped in favor of the
explicit container_of() expression, and cobalt_current_thread() shall
be used to determine the current TCB, which has to originates from the
cobalt/posix base personality.